# nodejs-client/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Rakam API Documentation",
    "description": "An analytics platform API that lets you create your own analytics services.",
    "version": "1.0.0"
  },
  "produces": ["application/json"],
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
...
  "/automation/remove": {
    
    "post": {
      "summary": "Remove rule",
      "description":"",
      "x-swagger-router-controller": "Automation",
      "tags": ["Automation"],
      "operationId": "removeRule",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "id",
  "in" : "formData",
  "required" : false,
  "type" : "integer",
  "format" : "int32"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        
        
      }
    } 
    
  } ,
  "/automation/activate": {
    
    "post": {
      "summary": "Activate rule",
      "description":"",
      "x-swagger-router-controller": "Automation",
      "tags": ["Automation"],
      "operationId": "activateRule",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "id",
  "in" : "formData",
  "required" : false,
  "type" : "integer",
  "format" : "int32"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        
        
      }
    } 
    
  } ,
  "/automation/deactivate": {
    
    "post": {
      "summary": "Deactivate rule",
      "description":"",
      "x-swagger-router-controller": "Automation",
      "tags": ["Automation"],
      "operationId": "deactivateRule",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
},
        {
  "name" : "id",
  "in" : "formData",
  "required" : false,
  "type" : "integer",
  "format" : "int32"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        
        
      }
    } 
    
  } ,
  "/automation/list": {
    
    "post": {
      "summary": "List scenarios",
      "description":"",
      "x-swagger-router-controller": "Automation",
      "tags": ["Automation"],
      "operationId": "listRules",
      "parameters": [
        {
  "name" : "project",
  "in" : "formData",
  "required" : false,
  "type" : "string"
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "type" : "array",
    "items" : {
      "$ref" : "#/definitions/AutomationRule"
    }
  }
}
        
        
      }
    } 
    
  } ,
  "/automation/add": {
    
    "post": {
      "summary": "Add scenario",
      "description":"",
      "x-swagger-router-controller": "Automation",
      "tags": ["Automation"],
      "operationId": "addRule",
      "parameters": [
        {
  "in" : "body",
  "name" : "AutomationRule",
  "required" : true,
  "schema" : {
    "$ref" : "#/definitions/AutomationRule"
  }
}
        
      ],
      "responses": {
        "200": {
  "description" : "successful operation",
  "schema" : {
    "$ref" : "#/definitions/JsonResponse"
  }
}
        
        
      }
    } 
    
  } ,
...
  }, "definitions": {
...
},"JsonResponse": {
  "properties" : {
    "success" : {
      "type" : "boolean",
      "default" : false
    },
    "message" : {
      "type" : "string"
    }
  }
},"Send": {
...
},"Threshold": {
  "properties" : {
    "aggregation" : {
      "type" : "string",
      "enum" : [ "count", "sum" ]
    },
    "fieldName" : {
      "type" : "string"
    },
    "value" : {
      "type" : "integer",
      "format" : "int64"
    }
  }
},"MaterializedViewBuilder": {
...
},"SerializableAction": {
  "properties" : {
    "type" : {
      "type" : "string",
      "enum" : [ "client", "user_action", "event" ]
    },
    "value" : {
      "type" : "object",
      "properties" : { }
    }
  }
},"RetentionAction": {
...
},"ScenarioStep": {
  "properties" : {
    "collection" : {
      "type" : "string"
    },
    "filter" : {
      "type" : "string"
    },
    "threshold" : {
      "$ref" : "#/definitions/Threshold"
    }
  }
},"Sorting": {
...
},"AutomationRule": {
  "required" : [ "actions", "project", "scenarios" ],
  "properties" : {
    "project" : {
      "type" : "string"
    },
    "is_active" : {
      "type" : "boolean",
      "default" : false
    },
    "scenarios" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/ScenarioStep"
      }
    },
    "actions" : {
      "type" : "array",
      "items" : {
        "$ref" : "#/definitions/SerializableAction"
      }
    },
    "custom_data" : {
      "type" : "string"
    },
    "id" : {
      "type" : "integer",
      "format" : "int32"
    }
  }
},"Variant": {
...
  }
}